requirements.txt
matplotlib==3.5.1
mediapipe==0.8.9.1
numpy==1.21.4
opencv-python-headless==4.5.5.64
Pillow==9.1.0
streamlit==1.5.0
tensorflow==2.8.0
click==8.0.4
scikit-image==0.19.2
scikit-learn==1.0.2
pandas==1.4.2
joblib==1.1.0
